code style

+++ kornia/augmentation/base.py
@@ -255,9 +255,9 @@ class _AugmentationBase(_BasicAugmentationBase):
         else:
             output = in_tensor.clone()
             trans_matrix = self.identity_matrix(in_tensor)
-            trans_matrix[to_apply] = self.compute_transformation(
-                in_tensor[to_apply], params=params, flags=flags
-            ).type(in_tensor.dtype)
+            trans_matrix[to_apply] = self.compute_transformation(in_tensor[to_apply], params=params, flags=flags).type(
+                in_tensor.dtype
+            )
             output[to_apply] = self.apply_transform(
                 in_tensor[to_apply], params=params, flags=flags, transform=trans_matrix[to_apply]
             ).type(in_tensor.dtype)
